Fix #10281: Growl/Messages refactor to UINotifications
