+++ primefaces/src/main/java/org/primefaces/component/message/MessageBase.java
@@ -23,25 +23,19 @@
  */
 package org.primefaces.component.message;
 
-import javax.faces.component.UIMessage;
-
-import org.primefaces.component.api.UINotification;
+import org.primefaces.component.api.UINotifications;
 import org.primefaces.component.api.Widget;
 
-public abstract class MessageBase extends UIMessage implements UINotification, Widget {
+public abstract class MessageBase extends UINotifications implements Widget {
 
     public static final String COMPONENT_FAMILY = "org.primefaces.component";
 
     public static final String DEFAULT_RENDERER = "org.primefaces.component.MessageRenderer";
 
     public enum PropertyKeys {
-
         display,
-        escape,
-        severity,
         style,
         styleClass,
-        skipDetailIfEqualsSummary
     }
 
     public MessageBase() {
@@ -61,23 +55,6 @@ public abstract class MessageBase extends UIMessage implements UINotification, W
         getStateHelper().put(PropertyKeys.display, display);
     }
 
-    public boolean isEscape() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.escape, true);
-    }
-
-    public void setEscape(boolean escape) {
-        getStateHelper().put(PropertyKeys.escape, escape);
-    }
-
-    @Override
-    public String getSeverity() {
-        return (String) getStateHelper().eval(PropertyKeys.severity, null);
-    }
-
-    public void setSeverity(String severity) {
-        getStateHelper().put(PropertyKeys.severity, severity);
-    }
-
     public String getStyle() {
         return (String) getStateHelper().eval(PropertyKeys.style, null);
     }
@@ -92,14 +69,5 @@ public abstract class MessageBase extends UIMessage implements UINotification, W
 
     public void setStyleClass(String styleClass) {
         getStateHelper().put(PropertyKeys.styleClass, styleClass);
-    }
-
-    @Override
-    public boolean isSkipDetailIfEqualsSummary() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.skipDetailIfEqualsSummary, false);
-    }
-
-    public void setSkipDetailIfEqualsSummary(boolean skipDetailIfEqualsSummary) {
-        getStateHelper().put(PropertyKeys.skipDetailIfEqualsSummary, skipDetailIfEqualsSummary);
     }
 }
